requirements.txt
flask==3.1.1
flask-cors
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0
pillow
accelerate==1.3.0
bitsandbytes==0.45.1
requests==2.32.3
numpy